1. Rename modules.
