+++ framework/cluster_manager/jvm/main/resources/cluster-manager.properties
@@ -1,0 +1,22 @@
+#
+# Copyright (c) 2019 - now, Eggroll Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+#
+#
+
+eggroll.cluster.manager.jdbc.driverClassName=com.mysql.cj.jdbc.Driver
+eggroll.cluster.manager.jdbc.url=jdbc:mysql://localhost:3306/eggroll_meta?useSSL=false&serverTimezone=UTC&characterEncoding=utf8&allowPublicKeyRetrieval=true
+eggroll.cluster.manager.jdbc.username=
+eggroll.cluster.manager.jdbc.password=
