+++ framework/cluster_manager/jvm/main/resources/eggroll-meta-mybatis-generator.xml
@@ -1,0 +1,107 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright (c) 2019 - now, Eggroll Authors. All Rights Reserved.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~
+  -->
+
+<!DOCTYPE generatorConfiguration
+        PUBLIC "-//mybatis.org//DTD MyBatis Generator Configuration 1.0//EN"
+        "http://mybatis.org/dtd/mybatis-generator-config_1_0.dtd">
+<generatorConfiguration>
+    <properties resource="cluster-manager.properties"/>
+    <!-- database driver-->
+    <context id="DB2Tables" targetRuntime="MyBatis3">
+        <property name="autoDelimitKeywords" value="true"/>
+        <property name="beginningDelimiter" value="`"/>
+        <property name="endingDelimiter" value="`"/>
+
+        <plugin type="org.mybatis.generator.plugins.CaseInsensitiveLikePlugin"/>
+        <plugin type="org.mybatis.generator.plugins.RowBoundsPlugin"/>
+        <plugin type="org.mybatis.generator.plugins.ToStringPlugin"/>
+        <plugin type="org.mybatis.generator.plugins.SerializablePlugin"/>
+        <plugin type="org.mybatis.generator.plugins.EqualsHashCodePlugin"/>
+        <plugin type="org.mybatis.generator.plugins.UnmergeableXmlMappersPlugin"/>
+
+        <commentGenerator>
+            <!-- whether to suppress generating date info -->
+            <property name="suppressDate" value="true"/>
+            <!-- whether to supress generatine comments -->
+            <property name="suppressAllComments" value="false"/>
+        </commentGenerator>
+
+        <!-- database connection info -->
+        <jdbcConnection driverClass="${eggroll.cluster.manager.jdbc.driverClassName}"
+                        connectionURL="${eggroll.cluster.manager.jdbc.url}"
+                        userId="${eggroll.cluster.manager.jdbc.username}"
+                        password="${eggroll.cluster.manager.jdbc.password}">
+        </jdbcConnection>
+        <javaTypeResolver type="org.mybatis.generator.internal.types.JavaTypeResolverDefaultImpl">
+            <!-- This property is used to specify whether MyBatis Generator should
+                force the use from java.math.BigDecimal for DECIMAL and NUMERIC fields, -->
+            <property name="forceBigDecimals" value="false"/>
+        </javaTypeResolver>
+        <!-- generated model package and location -->
+        <javaModelGenerator targetPackage="com.webank.eggroll.framework.clustermanager.dao.generated.model"
+                            targetProject="cluster_manager/jvm/src/main/java">
+            <property name="enableSubPackages" value="true"/>
+            <property name="trimStrings" value="true"/>
+            <property name="javaFileEncoding" value="UTF-8"/>
+            <property name="javaFormatter" value="org.mybatis.generator.api.dom.DefaultJavaFormatter"/>
+            <property name="xmlFormatter" value="org.mybatis.generator.api.dom.DefaultXmlFormatter"/>
+            <property name="constructorBased" value="false"/>
+            <property name="immutable" value="false"/>
+        </javaModelGenerator>
+        <!-- generated mapping package and location -->
+        <sqlMapGenerator targetPackage="com.webank.eggroll.framework.clustermanager.dao.generated.model"
+                         targetProject="cluster_manager/jvm/src/main/java">
+            <property name="enableSubPackages" value="true"/>
+            <property name="trimStrings" value="true"/>
+        </sqlMapGenerator>
+        <!-- generated DAO package and locations -->
+        <javaClientGenerator type="XMLMAPPER"
+                             targetPackage="com.webank.eggroll.framework.clustermanager.dao.generated.model"
+                             targetProject="${target.project}/src/main/java">
+            <property name="enableSubPackages" value="true"/>
+            <property name="trimStrings" value="true"/>
+        </javaClientGenerator>
+
+        <!-- tables to generate -->
+        <table schema="eggroll_meta" tableName="store" enableCountByExample="true"
+               enableDeleteByExample="true"
+               enableSelectByExample="true"
+               enableUpdateByExample="true"
+               selectByExampleQueryId="true">
+            <generatedKey column="store_id" sqlStatement="JDBC" identity="true"/>
+        </table>
+
+        <table schema="eggroll_meta" tableName="partition" enableCountByExample="true"
+               enableDeleteByExample="true"
+               enableSelectByExample="true"
+               enableUpdateByExample="true"
+               selectByExampleQueryId="true">
+            <generatedKey column="partition_id" sqlStatement="JDBC" identity="true"/>
+        </table>
+
+        <table schema="eggroll_meta" tableName="node" enableCountByExample="true"
+               enableDeleteByExample="true"
+               enableSelectByExample="true"
+               enableUpdateByExample="true"
+               selectByExampleQueryId="true">
+            <generatedKey column="node_id" sqlStatement="JDBC" identity="true"/>
+        </table>
+
+    </context>
+</generatorConfiguration>
