+++ framework/cluster_manager/jvm/main/resources/log4j2.properties
@@ -1,0 +1,107 @@
+#
+# Copyright (c) 2019 - now, Eggroll Authors. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+#
+#
+name=PropertiesConfig
+property.auditDir=audit
+property.logDir=logs
+property.project=eggroll
+property.module=cluster-manager
+property.logPattern=[%-5level] %d{yyyy-MM-dd}T%d{HH:mm:ss,SSS} [%t] [%c{1}:%L] - %msg%n
+# console
+appender.console.type=Console
+appender.console.name=STDOUT
+appender.console.layout.type=PatternLayout
+appender.console.layout.pattern=${logPattern}
+# default file
+appender.file.type=RollingFile
+appender.file.name=LOGFILE
+appender.file.fileName=${logDir}/${project}-${module}.log
+appender.file.filePattern=${logDir}/%d{yyyy}/%d{MM}/%d{dd}/${project}-${module}.log.%d{yyyy-MM-dd-HH}
+appender.file.layout.type=PatternLayout
+appender.file.layout.pattern=${logPattern}
+appender.file.policies.type=Policies
+appender.file.policies.time.type=TimeBasedTriggeringPolicy
+appender.file.policies.time.interval=1
+appender.file.policies.time.modulate=true
+appender.file.strategy.type=DefaultRolloverStrategy
+# debug
+appender.debugging.type=RollingFile
+appender.debugging.name=LOGDEBUGGING
+appender.debugging.fileName=${logDir}/${project}-${module}-debug.log
+appender.debugging.filePattern=${logDir}/%d{yyyy}/%d{MM}/%d{dd}/${project}-${module}-debug.log.%d{yyyy-MM-dd-HH-mm}
+appender.debugging.layout.type=PatternLayout
+appender.debugging.layout.pattern=${logPattern}
+appender.debugging.policies.type=Policies
+appender.debugging.policies.time.type=TimeBasedTriggeringPolicy
+appender.debugging.policies.time.interval=1
+appender.debugging.policies.time.modulate=true
+appender.debugging.strategy.type=DefaultRolloverStrategy
+# audit
+appender.audit.type=RollingFile
+appender.audit.name=LOGAUDIT
+appender.audit.fileName=${auditDir}/${project}-${module}-audit.log
+appender.audit.filePattern=${auditDir}/%d{yyyy}/%d{MM}/%d{dd}/${project}-${module}-audit.log.%d{yyyy-MM-dd-HH}
+appender.audit.layout.type=PatternLayout
+appender.audit.layout.pattern=[%d{yyyy-MM-dd}T%d{HH:mm:ss,SSS}]%msg%n
+appender.audit.policies.type=Policies
+appender.audit.policies.time.type=TimeBasedTriggeringPolicy
+appender.audit.policies.time.interval=1
+appender.audit.policies.time.modulate=true
+appender.audit.strategy.type=DefaultRolloverStrategy
+# stat
+appender.stat.type=RollingFile
+appender.stat.name=LOGSTAT
+appender.stat.fileName=${logDir}/${project}-${module}-stat.log
+appender.stat.filePattern=${logDir}/%d{yyyy}/%d{MM}/%d{dd}/${project}-${module}-stat.log.%d{yyyy-MM-dd-HH}
+appender.stat.layout.type=PatternLayout
+appender.stat.layout.pattern=${logPattern}
+appender.stat.policies.type=Policies
+appender.stat.policies.time.type=TimeBasedTriggeringPolicy
+appender.stat.policies.time.interval=1
+appender.stat.policies.time.modulate=true
+appender.stat.strategy.type=DefaultRolloverStrategy
+# loggers
+loggers=file, debugging, audit, stat
+# logger - file
+logger.file.name=file
+logger.file.level=info
+logger.file.appenderRefs=file
+logger.file.appenderRef.file.ref=LOGFILE
+logger.file.additivity=false
+# logger - debugging
+logger.debugging.name=debugging
+logger.debugging.level=info
+logger.debugging.appenderRefs=debugging
+logger.debugging.appenderRef.debugging.ref=LOGDEBUGGING
+logger.debugging.additivity=false
+# logger - audit
+logger.audit.name=audit
+logger.audit.level=info
+logger.audit.appenderRefs=audit
+logger.audit.appenderRef.file.ref=LOGAUDIT
+logger.audit.additivity=false
+# logger - stat
+logger.stat.name=stat
+logger.stat.level=info
+logger.stat.appenderRefs=stat
+logger.stat.appenderRef.file.ref=LOGSTAT
+logger.stat.additivity=false
+# logger - root
+rootLogger.level=info
+rootLogger.appenderRefs=stdout, file
+rootLogger.appenderRef.stdout.ref=STDOUT
+rootLogger.appenderRef.file.ref=LOGFILE
