+++ framework/cluster_manager/jvm/main/scala/com/webank/eggroll/cluster/manager/model/ClusterManagerModel.scala
