+++ framework/cluster_manager/jvm/pom.xml
@@ -1,0 +1,99 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright (c) 2019 - now, Eggroll Authors. All Rights Reserved.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~
+  -->
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xmlns="http://maven.apache.org/POM/4.0.0"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <artifactId>eggroll-cluster-manager</artifactId>
+  <modelVersion>4.0.0</modelVersion>
+
+  <name>Eggroll Cluster Manager</name>
+  <parent>
+    <artifactId>eggroll-all</artifactId>
+    <groupId>com.webank.eggroll</groupId>
+    <relativePath>../../../pom.xml</relativePath>
+    <version>${eggroll.version}</version>
+  </parent>
+
+  <properties>
+    <commons-dbcp2.version>2.7.0</commons-dbcp2.version>
+    <mybatis.version>3.5.3</mybatis.version>
+    <mybatis-generator-plugin.version>1.3.7</mybatis-generator-plugin.version>
+    <mysql-connector-java.version>8.0.18</mysql-connector-java.version>
+  </properties>
+
+  <dependencies>
+    <dependency>
+      <artifactId>eggroll-core</artifactId>
+      <groupId>com.webank.eggroll</groupId>
+      <version>${eggroll.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-dbcp2</artifactId>
+      <version>${commons-dbcp2.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.mybatis</groupId>
+      <artifactId>mybatis</artifactId>
+      <version>${mybatis.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>mysql</groupId>
+      <artifactId>mysql-connector-java</artifactId>
+      <version>${mysql-connector-java.version}</version>
+    </dependency>
+  </dependencies>
+
+  <build>
+    <resources>
+      <resource>
+        <directory>src/main/java</directory>
+        <includes>
+          <include>**/*.xml</include>
+        </includes>
+        <filtering>false</filtering>
+      </resource>
+    </resources>
+    <plugins>
+      <plugin>
+        <groupId>org.mybatis.generator</groupId>
+        <artifactId>mybatis-generator-maven-plugin</artifactId>
+        <version>${mybatis-generator-plugin.version}</version>
+        <dependencies>
+          <dependency>
+            <groupId>mysql</groupId>
+            <artifactId>mysql-connector-java</artifactId>
+            <version>${mysql-connector-java.version}</version>
+          </dependency>
+        </dependencies>
+        <configuration>
+          <configurationFile>${project.basedir}/src/main/resources/eggroll-meta-mybatis-generator.xml</configurationFile>
+          <overwrite>true</overwrite>
+          <verbose>true</verbose>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+
+
+
+
+
+
+</project>
