--- resource_managers/cluster_manager/jvm/pom.xml
@@ -1,74 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright (c) 2019 - now, Eggroll Authors. All Rights Reserved.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns="http://maven.apache.org/POM/4.0.0"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <artifactId>eggroll-cluster-manager</artifactId>
-
-  <build>
-    <plugins>
-      <plugin>
-        <artifactId>protobuf-maven-plugin</artifactId>
-        <configuration>
-          <additionalProtoPathElements>
-            <protoPathElement>${project.basedir}/../../../</protoPathElement>
-          </additionalProtoPathElements>
-          <pluginArtifact>io.grpc:protoc-gen-grpc-java:${grpc.version}:exe:${os.detected.classifier}
-          </pluginArtifact>
-          <pluginId>grpc-java</pluginId>
-          <protoSourceRoot>${pom.basedir}/../proto</protoSourceRoot>
-          <protocArtifact>
-            com.google.protobuf:protoc:${protobuf.version}:exe:${os.detected.classifier}
-          </protocArtifact>
-        </configuration>
-        <executions>
-          <execution>
-            <goals>
-              <goal>compile</goal>
-              <goal>compile-custom</goal>
-            </goals>
-          </execution>
-        </executions>
-        <groupId>org.xolstice.maven.plugins</groupId>
-        <version>${protobuf-maven-plugin.version}</version>
-      </plugin>
-    </plugins>
-  </build>
-
-  <dependencies>
-
-    <dependency>
-      <artifactId>eggroll-core</artifactId>
-      <groupId>com.webank.eggroll</groupId>
-      <version>${eggroll.version}</version>
-    </dependency>
-  </dependencies>
-  <modelVersion>4.0.0</modelVersion>
-
-  <name>Eggroll Cluster Manager</name>
-  <parent>
-    <artifactId>eggroll-all</artifactId>
-    <groupId>com.webank.eggroll</groupId>
-    <relativePath>../../../pom.xml</relativePath>
-    <version>${eggroll.version}</version>
-  </parent>
-
-  <properties>
-  </properties>
-
-
-</project>
